+++ Editor/AssetList/AssetListConfigurationInspector.cs
@@ -22,6 +22,8 @@ namespace Vertx.Editor
 
 		private SerializedProperty
 			assetType,
+			missingPropertyDisplay,
+			additionalColumns,
 			columns,
 			typeString,
 			nameDisplay,
@@ -32,6 +34,7 @@ namespace Vertx.Editor
 		private ReorderableList reorderableList;
 
 		private readonly GUIContent
+			preferencesLabel = new GUIContent("Preferences"),
 			titleLabel = new GUIContent("Title"),
 			iconLabel = new GUIContent("Icon"),
 			propertyLabel = new GUIContent("Property"),
@@ -41,13 +44,6 @@ namespace Vertx.Editor
 			cancelLabel = new GUIContent("Cancel"),
 			referenceObjectLabel = new GUIContent("Reference Object", "This object is used to gather Serialized Properties for column creation."),
 			searchlabel = new GUIContent("Serialized Property Search");
-
-		private GUIStyle centeredMiniLabel;
-
-		private GUIStyle CenteredMiniLabel => centeredMiniLabel ?? (centeredMiniLabel = new GUIStyle(EditorStyles.miniLabel)
-		{
-			alignment = TextAnchor.MiddleCenter
-		});
 
 		[SerializeField]
 		private AdvancedDropdownState propertyDropdownState;
@@ -102,6 +98,8 @@ namespace Vertx.Editor
 		{
 			base.OnEnable();
 			assetType = serializedObject.FindProperty("assetType");
+			missingPropertyDisplay = serializedObject.FindProperty("missingPropertyDisplay");
+			additionalColumns = serializedObject.FindProperty("additionalColumns");
 			columns = serializedObject.FindProperty("columns");
 			typeString = serializedObject.FindProperty("typeString");
 			nameDisplay = serializedObject.FindProperty("nameDisplay");
@@ -154,7 +152,7 @@ namespace Vertx.Editor
 						{
 							var arrayPropertyInformation = column.FindPropertyRelative("ArrayPropertyInformation");
 
-							ArrayDataDrawer.OnGUI(ref rect, propertyPath, arrayPropertyInformation, referenceObject, null);
+							ArrayDataDrawer.OnGUI(ref rect, propertyPath, arrayPropertyInformation, referenceObject, null, rect.x - 20, -16);
 
 							var arrayPropertyPath = arrayPropertyInformation.FindPropertyRelative("ArrayPropertyPath");
 							if (!string.IsNullOrEmpty(arrayPropertyPath.stringValue))
@@ -188,7 +186,7 @@ namespace Vertx.Editor
 				{
 					if (heightOverrideLookup.TryGetValue(index, out float height))
 						return height;
-					return EditorGUIExtensions.HeightWithSpacing * 3;
+					return EditorGUIUtils.HeightWithSpacing * 3;
 				},
 				onReorderCallback = list => heightOverrideLookup.Clear(),
 				displayAdd = false,
@@ -237,13 +235,22 @@ namespace Vertx.Editor
 				}
 			}
 
+			if (EditorGUIUtils.DrawHeaderWithFoldout(preferencesLabel, missingPropertyDisplay.isExpanded, headerXOffset: -16f))
+				missingPropertyDisplay.isExpanded = !missingPropertyDisplay.isExpanded;
+			if (missingPropertyDisplay.isExpanded)
+			{
+				EditorGUILayout.PropertyField(missingPropertyDisplay);
+				EditorGUILayout.PropertyField(additionalColumns);
+			}
+			EditorGUIUtils.DrawSplitter();
+
 			GUILayout.Space(8);
 
-			using (new EditorGUIExtensions.ContainerScope(columnsLabel, -2))
-			using (new EditorGUIExtensions.OutlineScope(false, false))
+			using (new EditorGUIUtils.ContainerScope(columnsLabel, -2))
+			using (new EditorGUIUtils.OutlineScope(false, false))
 			{
 				//Icon Header
-				GUILayout.Label(iconLabel, CenteredMiniLabel);
+				GUILayout.Label(iconLabel, EditorGUIUtils.CenteredMiniLabel);
 
 				if (typeIsTextureOrSprite)
 					EditorGUILayout.HelpBox("Type inherits from Texture or Sprite. Icon is automated.", MessageType.Info);
@@ -268,32 +275,36 @@ namespace Vertx.Editor
 						if (isArray)
 						{
 							float yBefore = rect.yMax;
-							ArrayDataDrawer.OnGUI(ref rect, iconPropertyPath, iconArrayPropertyInformation, referenceObject, typeStrings);
+							ArrayDataDrawer.OnGUI(ref rect, iconPropertyPath, iconArrayPropertyInformation, referenceObject, typeStrings, rect.x - 2);
 							GUILayoutUtility.GetRect(0, rect.yMax - yBefore);
 						}
 					}
 				}
 
-				GUILayout.Label(propertyLabel, CenteredMiniLabel);
+				GUILayout.Space(6);
+				EditorGUIUtils.DrawSplitter(alignXMinToZero: false);
+				GUILayout.Space(4);
+				GUILayout.Label(propertyLabel, EditorGUIUtils.CenteredMiniLabel);
 				using (new EditorGUI.DisabledScope(true))
-				{
 					EditorGUILayout.TextField(titleLabel, "Name");
-					EditorGUILayout.TextField("Property Path", "m_Name");
-				}
 
 				EditorGUILayout.PropertyField(nameDisplay);
 			}
 
 			reorderableList.DoLayoutList();
 
-			using (new EditorGUIExtensions.OutlineScope())
+			using (new EditorGUIUtils.OutlineScope())
 			{
 				GUILayout.Label(searchlabel, EditorStyles.boldLabel);
 				if (ValidateReferenceObjectWithHelpWarning(referenceObject))
 				{
 					Rect rect = EditorGUILayout.GetControlRect(false, EditorGUIUtility.singleLineHeight);
 					if (GUI.Button(rect, addLabel))
+					{
+						if (propertyDropdown == null)
+							CreatePropertyDropdown();
 						propertyDropdown.Show(rect);
+					}
 				}
 			}
 
@@ -315,7 +326,7 @@ namespace Vertx.Editor
 			if (referenceObject != null)
 				return true;
 			rect.NextGUIRect();
-			rect.height = EditorGUIExtensions.HeightWithSpacing * 2;
+			rect.height = EditorGUIUtils.HeightWithSpacing * 2;
 			EditorGUI.HelpBox(rect, referenceObjectMissingWarning, MessageType.Warning);
 			return false;
 		}
@@ -395,7 +406,6 @@ namespace Vertx.Editor
 
 				if (prop.propertyType == SerializedPropertyType.Generic && prop.isArray)
 				{
-					//TODO handle arrays. Check whether they have a relevant property in them.
 					bool hasTextureProperty = false;
 					SerializedProperty temp = prop.Copy();
 					SerializedProperty end = prop.GetEndProperty();